requirements.txt
numpy
matplotlib
warnings
argparse
sys
os
json
seaborn
pandas
scipy.stats
math
pprint
tempfile
typing
tensorflow
tensorflow_datasets
tensorflow_recommenders == 0.1.3
progressbar
time
collections
random
